+++ e2e/test/specs/js_sdk.spec.ts
@@ -4,7 +4,6 @@ import ChromeBrowserScreen from '../../src/screens/Android/ChromeBrowserScreen';
 import AndroidOpenWithComponent from '../../src/screens/Android/components/AndroidOpenWithComponent';
 import ReactNativeDappScreen from '../../src/screens/Dapps/ReactNativeDappScreen';
 import SdkPlaygroundDappScreen from '../../src/screens/Dapps/SdkPlaygroundDappScreen';
-import AndroidSDKDappScreen from '../../src/screens/Dapps/AndroidSDKDappScreen';
 import TestDappScreen from '../../src/screens/Dapps/TestDappScreen';
 import Web3OnBoardDappScreen from '../../src/screens/Dapps/Web3OnBoardDappScreen';
 import LockScreen from '../../src/screens/MetaMask/LockScreen';
@@ -19,7 +18,7 @@ import SafariBrowserScreen from '../../src/screens/iOS/SafariBrowserScreen';
 import IOSOpenInComponent from '../../src/screens/iOS/components/IOSOpenInComponent';
 import { beforeEachHook, beforeHook } from '../mocha.hooks';
 
-describe('JS SDK Connection', () => {
+describe('JS SDK E2E', () => {
   before(async () => {
     await beforeHook();
   });
@@ -146,10 +145,6 @@ describe('JS SDK Connection', () => {
 
     await SwitchNetworkModalComponent.switchNetwork();
 
-    if (driver.isAndroid) {
-      await SwitchNetworkModalComponent.switchNetwork();
-    }
-
     if (driver.isIOS) {
       await NetworkSwitchedModalComponent.tapGotItButton();
 
@@ -187,10 +182,6 @@ describe('JS SDK Connection', () => {
 
     await SignModalComponent.tapSignApproval();
 
-    if (driver.isAndroid) {
-      await SignModalComponent.tapSignApproval();
-    }
-
     if (driver.isIOS) {
       await driver.pause(1000);
       await Utils.launchApp(BROWSER_BUNDLE_ID);
@@ -206,8 +197,9 @@ describe('JS SDK Connection', () => {
 
     await SendTxModalComponent.reject();
 
+    await driver.pause(1000);
+
     if (driver.isIOS) {
-      await driver.pause(1000);
       await Utils.launchApp(BROWSER_BUNDLE_ID);
     }
   });
@@ -241,63 +233,6 @@ describe('JS SDK Connection', () => {
 
     await driver.pause(5000);
     await SignModalComponent.tapSignApproval();
-  });
-
-  it('Connect to the AndroidSDK Test Dapp', async () => {
-    await driver.pause(5000);
-
-    await Utils.launchApp(process.env.ANDROID_SDK_TEST_BUNDLE_ID ?? '');
-
-    await driver.pause(5000);
-
-    await AndroidSDKDappScreen.connect();
-
-    await driver.pause(3000);
-
-    await LockScreen.unlockMMifLocked(WALLET_PASSWORD);
-
-    await expect(
-      await ConnectModalComponent.connectApprovalButton,
-    ).toBeDisplayed();
-
-    await ConnectModalComponent.tapConnectApproval();
-
-    if (driver.isIOS) {
-      await driver.pause(1000);
-      await Utils.launchApp(process.env.ANDROID_SDK_TEST_BUNDLE_ID ?? '');
-    }
-
-    await driver.pause(5000);
-
-    await AndroidSDKDappScreen.sign();
-    await AndroidSDKDappScreen.sign2();
-
-    await SignModalComponent.tapSignApproval();
-
-    await AndroidSDKDappScreen.goBack();
-
-    await AndroidSDKDappScreen.batchSigning();
-    await AndroidSDKDappScreen.batchSigning();
-
-    await SignModalComponent.tapSignApproval();
-    await SignModalComponent.tapSignApproval();
-    await SignModalComponent.tapSignApproval();
-
-    await AndroidSDKDappScreen.goBack();
-
-    await AndroidSDKDappScreen.sendTransaction();
-    await AndroidSDKDappScreen.sendTransaction2();
-
-    await SendTxModalComponent.reject();
-
-    await AndroidSDKDappScreen.goBack();
-
-    await AndroidSDKDappScreen.switchChain();
-    await AndroidSDKDappScreen.switchChain2();
-
-    await SwitchNetworkModalComponent.switchNetwork();
-
-    await driver.pause(1000);
   });
 
   it.skip('Clear all connections', async () => {