+++ packages/sdk-communication-layer/src/config.ts
@@ -1,6 +1,6 @@
 export const DEFAULT_SERVER_URL =
   'https://metamask-sdk-socket.metafi.codefi.network/';
-export const DEFAULT_SOCKET_TRANSPORTS = ['websocket', 'polling'];
+export const DEFAULT_SOCKET_TRANSPORTS = ['websocket'];
 export const MIN_IN_MS = 1000 * 60;
 export const HOUR_IN_MS = MIN_IN_MS * 60;
 export const DAY_IN_MS = HOUR_IN_MS * 24;